added support to the down-converter (rtl-sdr driver)

+++ drivers/common/SequentialHopping.cpp
@@ -36,7 +36,7 @@ SequentialHopping::SequentialHopping() {
     }
 
 }
-int SequentialHopping::nextHop()
+long long  SequentialHopping::nextHop()
 {
     int ret_freq = mFreqs.at(mIndex);
     mIndex++;
